pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "setuptools-scm>=8.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "lakeshore-nomad-plugin"
version = "0.0.1"
description = "A plugin for NOMAD containing Lakeshore hall measurements."
readme = "README.md"
authors = [
    { name = "Andrea Albino", email = 'andrea.albino@physik.hu-berlin.de' },
    { name = "Hampus Näsström", email = 'hampus.naesstroem@physik.hu-berlin.de' },
    { name = "Sarthak Kapoor", email = 'sarthak.kapoor@ikz-berlin.de' },
    { name = "Sebastian Brückner", email = 'sebastian.brueckner@ikz-berlin.de' },
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    'nomad-lab',
    'nomad-material-processing',
    'nomad-measurements',
    'nomad-analysis',
]
[project.optional-dependencies]
dev = [
    "pytest",
    "structlog",
]

[project.license]
file = "LICENSE"

[project.urls]
"Homepage" = "https://github.com/FAIRmat-NFDI/nomad-material-processing"
"Bug Tracker" = "https://github.com/FAIRmat-NFDI/nomad-material-processing/issues"

[tool.setuptools.packages.find]
where = [
    "src",
]

[project.entry-points.'nomad.plugin']

schema = "lakeshore_nomad_plugin.hall:schema"
instrument_parser = "lakeshore_nomad_plugin.hall.instrument_parser:instrument_parser"
measurement_parser = "lakeshore_nomad_plugin.hall.measurement_parser:measurement_parser"

